requirements.txt
gradio
tensorflow-cpu
numpy
pillow